Add files via upload
This is the early version of the IPFS, a NFS. Most of the functions in this NFS can be found in the IPFS server.h and client.h files. This system works with 1 client and 1 server, and the server handles the files system. This server is able to handle multiple clients connecting to it.